+++ CHANGELOG.md
@@ -59,6 +59,8 @@ Main (unreleased)
 
 - Fix panic in `otelcol.receiver.syslog` when no tcp block was configured. (@kalleep)
 
+- Support Scrape Protocol specification in CRDS for `prometheus.operator.*` components. (@dehaansa) 
+
 - Support recovering from corrupted positions file entries in `loki.source.file`. (@dehaansa)
 
 v1.11.2